+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-# Changelog for HvcForge
+# Changelog for SSHForge
 
 ## v0.1.0 - TDB
 
-+ Initial version of the `HvcForge` module
++ Initial version of the `SSHForge` module